+++ Application/Features/Tenants/Dtos/UsageDto.cs
@@ -4,7 +4,7 @@
     {
         public string FeatureKey { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
-        public int Used { get; set; }
+        public long Used { get; set; }
         public int? Limit { get; set; }
         public string Unit { get; set; } = string.Empty;
         public bool? Enabled { get; set; }
